Use EditableEntityLabel in Longview instructions component
- Rename EditableTableRowLabel to EditableEntityLabel, since Longview is
no longer using a table layout
- Adjust the markup and stories in EditableEntityLabel (removed all Table
components)
- Use the updated component in LongviewClientInstructions, still with static
placeholder data
