+++ packages/manager/src/components/EditableEntityLabel/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './EditableEntityLabel';
